server/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-environ==0.11.2
channels[daphne]==4.1.0
django_cors_headers==4.6.0
channels_redis==4.2.0